chore: decompose dto & standardize names
