+++ android/src/main/java/com/giphyreactnativesdk/RNGiphySdkPackage.kt
@@ -7,21 +7,21 @@ import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ReactShadowNode
 import com.facebook.react.uimanager.ViewManager
 
-class GiphyReactNativeSdkPackage : ReactPackage {
+class RNGiphySdkPackage : ReactPackage {
   override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
     return listOf<NativeModule>(
       GiphySDKModule(reactContext),
-      GiphyReactNativeDialogModule(reactContext),
-      GiphyVideoManagerModule(reactContext),
+      RNGiphyDialogModule(reactContext),
+      RNGiphyVideoManagerModule(reactContext),
     )
   }
 
   override fun createViewManagers(reactContext: ReactApplicationContext):
     List<ViewManager<out View, out ReactShadowNode<*>>> {
     return listOf(
-      GiphyMediaViewManager(),
-      GiphyGridViewManager(),
-      GiphyVideoViewManager()
+      RNGiphyMediaViewManager(),
+      RNGiphyGridViewManager(),
+      RNGiphyVideoViewManager()
     )
   }
 }
