+++ android/src/main/java/com/giphyreactnativesdk/RNGiphyVideoManagerModule.kt
@@ -4,7 +4,7 @@ import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
 import com.facebook.react.bridge.ReactMethod
 
-class GiphyVideoManagerModule(reactContext: ReactApplicationContext) :
+class RNGiphyVideoManagerModule(reactContext: ReactApplicationContext) :
   ReactContextBaseJavaModule(reactContext) {
   override fun getName(): String {
     return "GiphyReactNativeVideoManager"
