+++ android/src/main/java/com/giphyreactnativesdk/dto/RNGiphyRendition.kt
@@ -1,0 +1,14 @@
+package com.giphyreactnativesdk.dto
+
+import com.giphy.sdk.core.models.enums.RenditionType
+import com.giphyreactnativesdk.utils.CaseConverter
+
+object RNGiphyRendition {
+  fun fromRNValue(renditionName: String?): RenditionType? {
+    return when (renditionName) {
+      null -> null
+      else -> RenditionType.values()
+        .firstOrNull { it.name == CaseConverter.snakeToCamel(renditionName) }
+    }
+  }
+}
