+++ src/dto/giphyContentType.ts
@@ -1,0 +1,8 @@
+export enum GiphyContentType {
+  Emoji = 'emoji',
+  Gif = 'gif',
+  Recents = 'recents',
+  Sticker = 'sticker',
+  Text = 'text',
+  Clips = 'clips',
+}
