+++ src/dto/giphyFileExtension.ts
@@ -1,0 +1,5 @@
+export enum GiphyFileExtension {
+  GIF = 'gif',
+  MP4 = 'mp4',
+  WebP = 'webp',
+}
