+++ src/dto/giphyMediaType.ts
@@ -1,0 +1,6 @@
+export enum GiphyMediaType {
+  Gif = 'gif',
+  Sticker = 'sticker',
+  Text = 'text',
+  Video = 'video',
+}
