+++ src/dto/giphyRating.ts
@@ -1,0 +1,8 @@
+export enum GiphyRating {
+  G = 'g',
+  PG = 'pg',
+  PG13 = 'pg-13',
+  R = 'r',
+  Unrated = 'unrated',
+  Y = 'y',
+}
