+++ src/dto/giphyRendition.ts
@@ -1,36 +1,5 @@
 import type { ConditionalKeys } from 'type-fest'
 
-export enum GiphyThemePreset {
-  Automatic = 'automatic',
-  Dark = 'dark',
-  Light = 'light',
-}
-
-export enum GiphyContentType {
-  Emoji = 'emoji',
-  Gif = 'gif',
-  Recents = 'recents',
-  Sticker = 'sticker',
-  Text = 'text',
-  Clips = 'clips',
-}
-
-export enum GiphyMediaType {
-  Gif = 'gif',
-  Sticker = 'sticker',
-  Text = 'text',
-  Video = 'video',
-}
-
-export enum GiphyRating {
-  G = 'g',
-  PG = 'pg',
-  PG13 = 'pg-13',
-  R = 'r',
-  Unrated = 'unrated',
-  Y = 'y',
-}
-
 export enum GiphyRendition {
   Original = 'original',
   OriginalStill = 'original_still',
@@ -80,43 +49,4 @@ export const GiphyClipsRendition: Record<
   FixedWidthDownsampled: GiphyRendition.FixedWidthDownsampled,
   DownsizedMedium: GiphyRendition.DownsizedMedium,
   DownsizedLarge: GiphyRendition.DownsizedLarge,
-}
-
-export enum GiphyStickersColumnCount {
-  Two = 2,
-  Three,
-  Four,
-}
-
-export enum GiphyFileExtension {
-  GIF = 'gif',
-  MP4 = 'mp4',
-  WebP = 'webp',
-}
-
-export enum GiphyDirection {
-  Horizontal = 'horizontal',
-  Vertical = 'vertical',
-}
-
-export enum GiphyContentRequestType {
-  Trending = 'trending',
-  Search = 'search',
-  Emoji = 'emoji',
-  Recents = 'recents',
-  Animate = 'animate',
-}
-
-export type GiphyContentRequest = {
-  mediaType: GiphyMediaType
-  rating?: GiphyRating
-  requestType: GiphyContentRequestType
-  searchQuery: string
-}
-
-export enum ResizeMode {
-  Center = 'center',
-  Contain = 'contain',
-  Cover = 'cover',
-  Stretch = 'stretch',
 }
