+++ src/dto/giphyThemePreset.ts
@@ -1,0 +1,5 @@
+export enum GiphyThemePreset {
+  Automatic = 'automatic',
+  Dark = 'dark',
+  Light = 'light',
+}
