+++ src/dto/misc.ts
@@ -1,0 +1,20 @@
+export type KeyboardAppearance = 'default' | 'dark' | 'light'
+export type IndicatorStyle = 'default' | 'dark' | 'light'
+
+export enum GiphyStickersColumnCount {
+  Two = 2,
+  Three,
+  Four,
+}
+
+export enum GiphyDirection {
+  Horizontal = 'horizontal',
+  Vertical = 'vertical',
+}
+
+export enum ResizeMode {
+  Center = 'center',
+  Contain = 'contain',
+  Cover = 'cover',
+  Stretch = 'stretch',
+}
