+++ src/native/GiphyMediaView.ts
@@ -1,7 +1,8 @@
 import { requireNativeComponent, ViewProps } from 'react-native'
 
-import type { GiphyMediaID } from '../giphyMedia'
-import type { GiphyRendition, ResizeMode } from './types'
+import type { GiphyMediaID } from '../dto/giphyMedia'
+import type { GiphyRendition } from '../dto/giphyRendition'
+import type { ResizeMode } from '../dto/misc'
 
 export const COMPONENT_NAME = 'GiphyReactNativeMediaView'
